% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sample.R
\name{slice_block}
\alias{slice_block}
\title{Sample complete groups from a data frame}
\usage{
slice_block(.data, ...)
}
\arguments{
\item{.data}{A data frame, possibly grouped}

\item{...}{Additional arguments passed to methods}

\item{n}{Number of samples to draw}

\item{replace}{Should sampling be done with replacement?}

\item{weight_by}{Optional expression to weight groups by (unquoted name)}
}
\value{
A data frame with sampled complete groups
}
\description{
Sample complete groups from a data frame
}
